WaiterOverlay: move it to from silx.gui.utils to silx.gui.widgets
