+++ src/silx/gui/widgets/waiteroverlay.py
